desktop/requirements.txt
Pillow==11.1.0
pytesseract==0.3.13
python-dotenv==1.0.1
pynput==1.7.7
requests==2.32.3
